fix(cli): Refactor cli and add modifier chain for modifications.
