+++ packages/cli/src/client/typescript-converter.test.ts
@@ -1,0 +1,4 @@
+
+test('Generate typescript from json-schema', () => {
+
+})
